Move tcpconnect cmd construction to cmd/common
This refactor prepares for creating a local-gadget version
of the tcpconnect cmd.

Related to: https://github.com/kinvolk/inspektor-gadget/issues/895
